Rename components from Color to Shade to be more accurate
