+++ src/components/molecules/ShadeDisplay.vue
